Restore resampler.

+++ core/thread.h
@@ -84,7 +84,7 @@ struct Semaphore {
 		pthread_cond_signal(&condition);
 	}
 	/// Returns available ressources \a count
-	operator int() const { return counter; }
+	operator uint64() const { return counter; }
 };
 inline String str(const Semaphore& o) { return str(o.counter); }
